+++ seapopym/function/generator/__init__.py
@@ -2,8 +2,7 @@ from .apply_coefficient_to_primary_production import apply_coefficient_to_primar
 from .average_temperature import average_temperature_kernel
 from .cell_area import cell_area_kernel
 from .day_length import day_length_kernel
-from .global_mask import global_mask_kernel
-from .mask_by_fgroup import mask_by_fgroup_kernel
+from .mask import global_mask_kernel, mask_by_fgroup_kernel
 from .mask_temperature import mask_temperature_kernel
 from .min_temperature import min_temperature_kernel
 from .mortality_field import mortality_field_kernel
