+++ seapopym/function/generator/mask_temperature.py
@@ -1,15 +1,18 @@
 """A temperature mask computation wrapper. Use xarray.map_block."""
 from __future__ import annotations
 
+from typing import TYPE_CHECKING
+
 import cf_xarray  # noqa: F401
-import xarray as xr
 
 from seapopym.function.core.kernel import KernelUnits
 from seapopym.function.core.template import ForcingTemplate
 from seapopym.standard.attributs import mask_temperature_desc
 from seapopym.standard.labels import CoordinatesLabels, PreproductionLabels
-from seapopym.standard.types import SeapopymForcing
 from seapopym.standard.units import StandardUnitsLabels, check_units
+
+if TYPE_CHECKING:
+    import xarray as xr
 
 
 def _mask_temperature_helper(state: xr.Dataset) -> xr.DataArray:
@@ -41,27 +44,6 @@ def _mask_temperature_helper(state: xr.Dataset) -> xr.DataArray:
     return average_temperature >= min_temperature
 
 
-# def mask_temperature(state: xr.Dataset, chunk: dict | None = None, lazy: ForcingName | None = None) -> SeapopymForcing:
-#     """Wrap the average temperature by functional group computation with a map_block function."""
-#     class_type = Template if lazy is None else TemplateLazy
-#     template_attributs = {
-#         "name": PreproductionLabels.mask_temperature,
-#         "dims": [
-#             CoordinatesLabels.functional_group,
-#             CoordinatesLabels.time,
-#             CoordinatesLabels.Y,
-#             CoordinatesLabels.X,
-#             CoordinatesLabels.cohort,
-#         ],
-#         "attributs": mask_temperature_desc,
-#         "chunk": chunk,
-#     }
-#     if lazy is not None:
-#         template_attributs["model_name"] = lazy
-#     template = class_type(**template_attributs)
-#     return apply_map_block(function=_mask_temperature_helper, state=state, template=template)
-
-
 def mask_temperature_template(chunk: dict | None = None) -> ForcingTemplate:
     return ForcingTemplate(
         name=PreproductionLabels.mask_temperature,
@@ -77,7 +59,7 @@ def mask_temperature_template(chunk: dict | None = None) -> ForcingTemplate:
     )
 
 
-def mask_temperature_kernel(*, chunk: dict | None = None, template: ForcingTemplate | None = None) -> SeapopymForcing:
+def mask_temperature_kernel(*, chunk: dict | None = None, template: ForcingTemplate | None = None) -> KernelUnits:
     if template is None:
         template = mask_temperature_template(chunk=chunk)
     return KernelUnits(
