+++ seapopym/function/generator/min_temperature.py
@@ -9,7 +9,6 @@ from seapopym.function.core.kernel import KernelUnits
 from seapopym.function.core.template import ForcingTemplate
 from seapopym.standard.attributs import min_temperature_by_cohort_desc
 from seapopym.standard.labels import ConfigurationLabels, CoordinatesLabels, PreproductionLabels
-from seapopym.standard.types import SeapopymForcing
 
 
 def _min_temperature_by_cohort_helper(state: xr.Dataset) -> xr.DataArray:
@@ -33,22 +32,6 @@ def _min_temperature_by_cohort_helper(state: xr.Dataset) -> xr.DataArray:
     )
 
 
-# def min_temperature(state: xr.Dataset, chunk: dict | None = None, lazy: ForcingName | None = None) -> SeapopymForcing:
-#     """Wrap the average temperature by functional group computation with a map_block function."""
-#     class_type = Template if lazy is None else TemplateLazy
-#     template_attributs = {
-#         "name": PreproductionLabels.min_temperature,
-#         "dims": [CoordinatesLabels.functional_group, CoordinatesLabels.cohort],
-#         "attributs": min_temperature_by_cohort_desc,
-#         "chunk": chunk,
-#     }
-#     if lazy is not None:
-#         template_attributs["model_name"] = lazy
-#     template = class_type(**template_attributs)
-
-#     return apply_map_block(function=_min_temperature_by_cohort_helper, state=state, template=template)
-
-
 def min_temperature_template(chunk: dict | None = None) -> ForcingTemplate:
     return ForcingTemplate(
         name=PreproductionLabels.min_temperature,
@@ -58,7 +41,7 @@ def min_temperature_template(chunk: dict | None = None) -> ForcingTemplate:
     )
 
 
-def min_temperature_kernel(*, chunk: dict | None = None, template: ForcingTemplate | None = None) -> SeapopymForcing:
+def min_temperature_kernel(*, chunk: dict | None = None, template: ForcingTemplate | None = None) -> KernelUnits:
     if template is None:
         template = min_temperature_template(chunk=chunk)
     return KernelUnits(
