+++ test/unit/function/generator/test_production.py
@@ -2,14 +2,15 @@ import cf_xarray  # noqa: F401
 import numpy as np
 import xarray as xr
 
-from seapopym.function.generator.production.production import production
+from seapopym.function.generator.production.production import production_kernel
 from seapopym.standard.labels import CoordinatesLabels, ProductionLabels
 
 
 class TestProduction:
     def test_mask_by_fgroup_without_init_export_chunk(self, state_production_fg4_t4d_y1_x1_c4):
         state_production_fg4_t4d_y1_x1_c4 = state_production_fg4_t4d_y1_x1_c4.drop_vars("initial_condition_production")
-        results = production(state_production_fg4_t4d_y1_x1_c4, export_preproduction=None)
+        kernel = production_kernel()
+        results = kernel.run(state_production_fg4_t4d_y1_x1_c4)
         results = results[ProductionLabels.recruited]
         assert isinstance(results, xr.DataArray)
         dims = (
@@ -33,8 +34,8 @@ class TestProduction:
         chunk = {"Y": 1, "X": 1}
         state_production_fg4_t4d_y1_x1_c4 = state_production_fg4_t4d_y1_x1_c4.drop_vars("initial_condition_production")
         state_production_fg4_t4d_y1_x1_c4 = state_production_fg4_t4d_y1_x1_c4.cf.chunk(chunk)
-
-        results = production(state_production_fg4_t4d_y1_x1_c4, chunk=chunk, export_preproduction=None).compute()
+        kernel = production_kernel(chunk=chunk)
+        results = kernel.run(state_production_fg4_t4d_y1_x1_c4).compute()
         results = results[ProductionLabels.recruited]
         assert isinstance(results, xr.DataArray)
         dims = (
